chainlit/requirements.txt
chainlit>=0.7.0
langchain>=0.1.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
openai>=1.0.0

pyodbc
azure-identity

langchain-community 
langchain-sqlserver
langgraph

chainlit


